backend/engine/templates/generated/a3633e1b-0ef9-439d-a79c-972d8caac0fd/requirements.txt
fastmcp>=0.1.0
pydantic>=1.10,<2.0
# OR if using Pydantic v2:
# pydantic>=2.0
python-dotenv>=1.0.0
# Specify the correct version of pykiteconnect you intend to use
pykiteconnect>=3.9.0,<4.0
uvicorn>=0.20.0
# httpx is used by fastmcp and potentially pykiteconnect internally
httpx>=0.23.0
